backend/requirements.txt
fastapi
uvicorn
opencv-python
numpy
tqdm
torch
pytorch
ffmpeg-python